Restructure DirectExecutorV2 codegen and externs module so that more codgen code and externs can be shared with StateExecutor. Also update the microcode executor to share its externs with DirectExecutorV2, at least for the externs that aren't control-flow-modifying.
