+++ feo3boy-executor-generator/src/lib.rs
@@ -3,25 +3,21 @@
 use proc_macro2::{Ident, Span, TokenStream};
 use proc_macro_crate::FoundCrate;
 use quote::quote;
-use syn::{parse_macro_input, ItemMod};
+use syn::parse_macro_input;
 
 use crate::direct_executor::DirectExecutor;
+use crate::parsing::ExecutorDef;
 
 mod direct_executor;
+mod parsing;
 
 /// Generates an implementation of the DirectExecutor.
-#[proc_macro_attribute]
-pub fn define_direct_executor(
-    attr: proc_macro::TokenStream,
-    item: proc_macro::TokenStream,
-) -> proc_macro::TokenStream {
-    // Name to use for the executor.
-    let name = parse_macro_input!(attr as Ident);
+#[proc_macro]
+pub fn define_direct_executor(item: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    // Name and mapping of externs for the executor.
+    let executor = parse_macro_input!(item as ExecutorDef);
 
-    // Contents of the module which defines externs.
-    let module = parse_macro_input!(item as ItemMod);
-
-    let def = match DirectExecutor::extract(name, module) {
+    let def = match DirectExecutor::extract(executor) {
         Ok(def) => def,
         Err(e) => return e.into_compile_error().into(),
     };
