+++ feo3boy-opcodes/src/compiler/mod.rs
@@ -6,6 +6,7 @@ use proc_macro2::TokenStream;
 pub mod args;
 pub mod direct_executor_generation;
 pub mod instr;
+pub mod variables;
 
 /// Which type of microcode operation this is.
 #[derive(Debug, Clone)]
